requirements.txt
# anaconda
elasticsearch
lxml
requests

# pip
readability-lxml
